tests/requirements.txt
django_nose
markdown
https://github.com/InfoAgeTech/django-core/tarball/master
https://github.com/InfoAgeTech/django-testing/tarball/master